# openapi.yaml
openapi: 3.0.0
info:
    title: IbyeolNote
    version: 1.0.0
    description: API documentation for IbyeolNote
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    schemas:
...
        PostResponse:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                    example: d290f1ee-6c54-4b01-90e6-d701748f0851
                content:
                    type: string
                    example: 게시글 내용
                view:
                    type: number
                    example: 조회수
                createdAt:
                    type: string
                    format: date-time
                    example: "2024-10-22T09:00:00Z"
                updatedAt:
                    type: string
                    format: date-time
                    nullable: true
                    example: "2024-10-23T09:00:00Z"
                isDeleted:
                    type: boolean
                    example: false
                deletedAt:
                    type: string
                    format: date-time
                    nullable: true
                    example: null
                RelationshipDecision:
                    type: string
                    enum:
                        - HOLD_ON
                        - LET_GO
                    example: LET_GO
                user:
                    type: object
                    properties:
                        id:
                            type: number
                        name:
                            type: string
                        image:
                            type: string
                            format: url
                        isBookmark:
                            type: boolean
        CreatePostDto:
            type: object
            required:
                - content
            properties:
                content:
                    type: string
                    example: 새로운 게시글 내용
        UpdatePostDto:
            type: object
            properties:
                content:
                    type: string
                    example: 수정된 게시글 내용
...
        CommentResponse:
            type: object
            properties:
              id:
                  type: string
                  format: uuid
                  example: b123f1ee-7c45-5d01-91a6-c801748f0451
              content:
                  type: string
                  example: 댓글 내용
              createdAt:
                  type: string
                  format: date-time
                  example: "2024-10-22T10:00:00Z"
              updatedAt:
                  type: string
                  format: date-time
                  nullable: true
                  example: "2024-10-23T10:00:00Z"
              isDeleted:
                  type: boolean
                  example: false
              deletedAt:
                  type: string
                  format: date-time
                  nullable: true
                  example: null
              user:
                  type: object
                  properties:
                      id:
                          type: string
                          format: uuid
                          example: a123f1ee-8d54-5f01-92e7-e801748f0851
                      name:
                          type: string
                          example: "username123"
paths:
...
    /post:
        get:
            summary: 게시글 목록 조회
            tags:
                - Posts
            parameters:
                - name: cursor
                  in: query
                  required: false
                  schema:
                      type: string
                      description: 현재 페이지에서 마지막 게시글의 ID
                - name: limit
                  in: query
                  required: false
                  schema:
                      type: integer
                      default: 5
                      description: 페이지당 게시글 수
            responses:
                "200":
                    description: 게시글 목록 조회 성공
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/PostResponse"
        post:
            summary: 새 게시글 작성
            tags:
                - Posts
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/CreatePostDto"
            responses:
                "201":
                    description: 게시글 작성 성공
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PostResponse"
    /post/{id}:
        get:
            summary: 특정 게시글 조회
            tags:
                - Posts
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
            responses:
                "200":
                    description: 게시글 조회 성공
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PostResponse"
                "404":
                    description: 게시글을 찾을 수 없음
        put:
            summary: 기존 게시글 수정
            tags:
                - Posts
            security:
                - bearerAuth: []
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UpdatePostDto"
            responses:
                "200":
                    description: 게시글 수정 성공
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PostResponse"
                "404":
                    description: 게시글을 찾을 수 없음
        delete:
            summary: 특정 게시글 삭제
            tags:
                - Posts
            security:
                - bearerAuth: []
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
            responses:
                "204":
                    description: 게시글 삭제 성공
                "404":
                    description: 게시글을 찾을 수 없음
    # 댓글 관련은 추후에 또 작성할게요!
    # /post/comment/{postId}:
    #     get:
    #         summary: 게시글에 대한 댓글 조회
    #         tags:
    #             - Post
    #         security:
    #             - bearerAuth: []
    #         parameters:
    #             - name: postId
    #               in: path
    #               required: true
    #               schema:
    #                   type: string
    #                   format: uuid
    #               description: 게시글 ID를 통한 댓글 조회
    #         responses:
    #             "200":
    #                 description: 댓글 조회 성공
    #                 content:
    #                     application/json:
    #                         schema:
    #                             $ref: "#/components/schemas/CommentResponse"
    #             "400":
    #                 description: postId 조회 실패
...